Fixed: problem with empty json user-file. Moved event messaged from ReflectorClient to Reflector class. More info in log file about user, state and Sds's
